+++ CHANGELOG.md
@@ -1,6 +1,22 @@
 <!-- Keep a Changelog guide -> https://keepachangelog.com -->
 
-# generate-flutter-mocks-plugin Changelog
+# Generate flutter mocks Changelog
+
+## [1.0.0] - 2024-01-11
+### Added
+- Default shortcut
+- Progress indicator
+
+### Changed
+- Generation method no longer blocks the main thread
+
+## [0.4.1] - 2024-01-11
+### Added
+- Add additional info to output popup
+
+## [0.4.0] - 2024-01-11
+### Added
+- Option to generate mocks
 
 ## [Unreleased]
 ### Added